+++ packages/integration-boilerplate-node/__tests__/jest-cucumber.ts
@@ -14,8 +14,9 @@ jest.useFakeTimers();
   "SAASQUATCH_AUTH0_SECRET",
   "SAASQUATCH_AUTH0_DOMAIN",
 ].forEach((e) => (process.env[e] = e));
-process.env["SERVER_LOG_LEVEL"] = "none";
+process.env["SERVER_LOG_LEVEL"] = "debug";
 process.env["ENFORCE_HTTPS"] = "false";
+process.env["SERVICE_NAME"] = "jest-tests";
 process.env["SAASQUATCH_APP_DOMAIN"] = "mocked_saasquatch";
 
 const features = loadFeatures("__tests__/features/**/*.feature");