+++ packages/integration-boilerplate-node/src/config.ts
@@ -44,6 +44,9 @@ export class BaseConfig {
   @optional("")
   public proxyFrontend!: string;
 
+  @key("SERVICE_NAME")
+  public serviceName!: string;
+
   @key("SAASQUATCH_APP_DOMAIN")
   @optional("app.referralsaasquatch.com")
   public saasquatchAppDomain!: string;
@@ -64,6 +67,14 @@ export class BaseConfig {
   @key("STATIC_FRONTEND_INDEX")
   @optional("index.html")
   public staticFrontendIndex!: string;
+
+  @key("INTROSPECTION_ENDPOINT_PATH")
+  @optional("/introspection")
+  public introspectionEndpointPath!: string;
+
+  @key("METRICS_ENABLED")
+  @optional(false)
+  public metricsEnabled!: boolean;
 }
 
 export async function loadConfig<T extends BaseConfig>(