+++ packages/integration-boilerplate-node/src/service.ts
@@ -1,25 +1,35 @@
-import path from "path";
-import express, { Express, Router, Response } from "express";
+import { httpLogMiddleware } from "@saasquatch/logger";
+import { hostname } from "os";
+import { IntegrationConfiguration } from "@saasquatch/schema/types/IntegrationConfig";
 import compression from "compression";
+import express, { Express, Response, Router } from "express";
 import enforce from "express-sslify";
 import { createProxyMiddleware } from "http-proxy-middleware";
-import { Logger } from "winston";
-import fetch from "node-fetch";
 import NodeCache from "node-cache";
-
-import * as types from "./types";
-import { BaseConfig, loadConfig } from "./config";
-import { createLogger } from "./logger";
+import fetch from "node-fetch";
+import path from "path";
+import { Logger } from "winston";
+import { gql } from ".";
 import { Auth } from "./auth";
+import { BaseConfig, loadConfig } from "./config";
+import { GraphQLError, IntegrationConfigError } from "./errors";
+import { formHandler } from "./formHandler";
+import { introspectionHandler } from "./introspectionHandler";
+import { createLogger } from "./logger";
+import {
+  MetricsManager,
+  METRIC_FORM_COUNT,
+  METRIC_INTROSPECTION_COUNT,
+  METRIC_REQUEST_PROCESSING_COUNT,
+  METRIC_WEBHOOK_COUNT,
+} from "./metrics";
 import {
   createSaasquatchRequestMiddleware,
   createSaasquatchTokenMiddleware,
 } from "./middleware";
+import * as types from "./types";
 import { webhookHandler } from "./webhookHandler";
-import { formHandler } from "./formHandler";
-import { IntegrationConfigError, GraphQLError } from "./errors";
-import { gql } from ".";
-import { httpLogMiddleware } from "@saasquatch/logger";
+import { installInstrumentation } from "./instrumentation";
 
 declare module "http" {
   interface IncomingMessage {
@@ -62,6 +72,12 @@ export interface IntegrationHandlers<
     context: types.FormInitialDataRequestContext<IntegrationConfig, FormConfig>,
     graphql: types.TenantScopedGraphQLFn
   ) => Promise<types.FormInitialDataResponse | types.FormErrorResponse>;
+  introspectionHandler?: (
+    service: IntegrationService<ServiceConfig, IntegrationConfig, FormConfig>,
+    config: IntegrationConfig,
+    templateIntegrationConfig: IntegrationConfiguration,
+    tenantAlias: types.TenantAlias
+  ) => Promise<types.IntrospectionResponse>;
 }
 
 export interface ServiceOptions<
@@ -81,6 +97,7 @@ export class IntegrationService<
 > {
   readonly config: ServiceConfig;
   readonly logger: Logger;
+  readonly metricsManager: MetricsManager | undefined;
   readonly auth: Auth;
   readonly options?: ServiceOptions<
     ServiceConfig,
@@ -102,6 +119,9 @@ export class IntegrationService<
     this.options = options;
     this.config = config;
     this.logger = createLogger(config);
+    this.metricsManager = config.metricsEnabled
+      ? new MetricsManager(config.serviceName)
+      : undefined;
     this.auth = new Auth(
       config.saasquatchAppDomain,
       config.saasquatchAuth0ClientId,
@@ -141,7 +161,7 @@ export class IntegrationService<
     `;
 
     interface TenantQueryData {
-      data: { viewer: { tenants: { tenantAlias: string }[] } };
+      data: { viewer: { tenants: { tenantAlias: types.TenantAlias }[] } };
     }
 
     const result = await this.graphql<TenantQueryData>(tenantQuery);
@@ -152,17 +172,19 @@ export class IntegrationService<
     return tenantAliases;
   }
 
-  async getTenant(tenantAlias: string) {
+  async getTenant(tenantAlias: types.TenantAlias) {
     const config = await this.getIntegrationConfig(tenantAlias);
     const graphql = this.getTenantScopedGraphQL(tenantAlias);
     return { config, graphql };
   }
 
-  async getUserGraphQL(tenantAlias: string, userJwt: string) {
+  async getUserGraphQL(tenantAlias: types.TenantAlias, userJwt: string) {
     return this.getUserScopedGraphQL(tenantAlias, userJwt);
   }
 
-  async getIntegrationConfig(tenantAlias: string): Promise<IntegrationConfig> {
+  async getIntegrationConfig(
+    tenantAlias: types.TenantAlias
+  ): Promise<IntegrationConfig> {
     if (this.tenantIntegrationConfigCache.has(tenantAlias)) {
       this.logger.debug(
         `Retrieving integration config for tenant [${tenantAlias}] from cache`
@@ -208,7 +230,7 @@ export class IntegrationService<
   }
 
   private getTenantScopedGraphQL(
-    tenantAlias: string
+    tenantAlias: types.TenantAlias
   ): types.TenantScopedGraphQLFn {
     return <QueryResponseShape>(
       query: string,
@@ -224,7 +246,7 @@ export class IntegrationService<
   }
 
   private getUserScopedGraphQL(
-    tenantAlias: string,
+    tenantAlias: types.TenantAlias,
     userJwt: string
   ): types.TenantScopedGraphQLFn {
     return <QueryResponseShape>(
@@ -244,7 +266,7 @@ export class IntegrationService<
   private async graphql<QueryResponseShape = unknown>(
     query: string,
     opts?: {
-      tenantAlias?: string;
+      tenantAlias?: types.TenantAlias;
       variables?: Record<string, any>;
       operationName?: string;
       token?: string;
@@ -314,6 +336,30 @@ export class IntegrationService<
     // Enable request logging
     server.use(httpLogMiddleware(this.logger));
 
+    // Record the server's HTTP response times if metrics are enabled
+    if (this.config.metricsEnabled) {
+      server.use(async (_req, res, next) => {
+        const startTimeNs = process.hrtime.bigint();
+        this.metricsManager!.increment(METRIC_REQUEST_PROCESSING_COUNT);
+
+        res.on("finish", () => {
+          const endTimeNs = process.hrtime.bigint();
+          const time = Number((endTimeNs - startTimeNs) / BigInt(1000));
+
+          if (!Number.isNaN(time) && Number.isFinite(time) && time > 0) {
+            this.metricsManager!.recordHistogramVal(
+              "response_time",
+              Number(time)
+            );
+          }
+
+          this.metricsManager!.decrement(METRIC_REQUEST_PROCESSING_COUNT);
+        });
+
+        next();
+      });
+    }
+
     // Force HTTPS for all requests
     if (this.config.enforceHttps) {
       server.use(enforce.HTTPS({ trustProtoHeader: true }));
@@ -346,7 +392,24 @@ export class IntegrationService<
           this,
           this.getTenantScopedGraphQL(req.body.tenantAlias)
         );
+
+        if (this.config.metricsEnabled) {
+          this.metricsManager!.increment(METRIC_WEBHOOK_COUNT);
+        }
       });
+    }
+
+    if (this.options?.handlers?.introspectionHandler) {
+      server.post(
+        this.config.introspectionEndpointPath,
+        requireSaaSquatchSignature,
+        async (req, res) => {
+          await introspectionHandler(req, res, this);
+          if (this.config.metricsEnabled) {
+            this.metricsManager!.increment(METRIC_INTROSPECTION_COUNT);
+          }
+        }
+      );
     }
 
     if (
@@ -362,6 +425,10 @@ export class IntegrationService<
           this,
           this.getTenantScopedGraphQL(req.body.tenantAlias)
         );
+
+        if (this.config.metricsEnabled) {
+          this.metricsManager!.increment(METRIC_FORM_COUNT);
+        }
       });
     }
 
@@ -413,5 +480,13 @@ export async function createIntegrationService<
   const config: ServiceConfig = options?.configClass
     ? await loadConfig(options.configClass)
     : await loadConfig();
+
+  if (config.metricsEnabled) {
+    const hostName = `${config.serviceName}.${
+      process.env["DYNO"] ?? hostname()
+    }`;
+    installInstrumentation(config.serviceName, hostName);
+  }
+
   return new IntegrationService(config, options);
 }